move `base_only` impl into its own module
